+++ kite/src/test/kotlin/com/tang/kite/session/mapper/AccountAsMapper.kt
@@ -1,0 +1,10 @@
+package com.tang.kite.session.mapper
+
+import com.tang.kite.mapper.BaseMapper
+import com.tang.kite.session.entity.AccountAs
+
+/**
+ * @author Tang
+ */
+interface AccountAsMapper : BaseMapper<AccountAs> {
+}